move review decision to types
